refactor lib files into hooks and utils
