+++ packages/lib/package.json
@@ -11,8 +11,9 @@
   "devDependencies": {
     "@types/react": "17.0.39",
     "@types/react-dom": "17.0.11",
-    "tsconfig": "*",
     "config": "*",
-    "typescript": "4.5.5"
+    "tsconfig": "*",
+    "tsup": "5.11.13",
+    "typescript": "4.6.2"
   }
 }
