--- packages/lib/src/includes.ts
@@ -1,5 +1,0 @@
-function includes<T> (list: T[], value: T) {
-  return list.indexOf(value) > -1
-}
-
-export default includes
