+++ packages/lib/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "tsconfig/base.json",
-  "include": ["src"],
+  "include": ["."],
   "exclude": ["dist", "build", "node_modules"]
 }
