+++ packages/lib/utils/array.ts
@@ -1,0 +1,3 @@
+export function includes<T> (list: T[], value: T) {
+  return list.indexOf(value) > -1;
+}
