+++ packages/lib/utils/date.ts
@@ -1,4 +1,4 @@
-export default function getCurrentTime () {
+export function getCurrentTime () {
   const date = new Date()
   const hours = date.getHours()
   const minutes = date.getMinutes()
