+++ packages/lib/utils/math.ts
