+++ packages/lib/utils/string.ts
@@ -1,4 +1,4 @@
-export const classNames = (...args: (string | { [key: string]: boolean })[]) => {
+export function classNames (...args: (string | { [key: string]: boolean })[]) {
   const classes: string[] = []
   for (const arg of args) {
     if (typeof arg === 'string') {
